+++ src/main/java/insane96mcp/survivalreimagined/mixin/LeavesBlockMixin.java
@@ -1,9 +1,7 @@
 package insane96mcp.survivalreimagined.mixin;
 
-import insane96mcp.insanelib.base.Feature;
 import insane96mcp.survivalreimagined.module.world.feature.FastLeafDecay;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.block.LeavesBlock;
@@ -17,16 +15,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class LeavesBlockMixin {
 	@Inject(method = "randomTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/level/ServerLevel;removeBlock(Lnet/minecraft/core/BlockPos;Z)Z", shift = At.Shift.AFTER))
 	private void onRandomTick(BlockState state, ServerLevel level, BlockPos pos, RandomSource randomSource, CallbackInfo ci) {
-		if (!Feature.isEnabled(FastLeafDecay.class))
-			return;
-
-		Direction.stream().forEach(direction -> {
-			BlockState adjacentState = level.getBlockState(pos.relative(direction));
-			if (!(adjacentState.getBlock() instanceof LeavesBlock))
-				return;
-
-			if (randomSource.nextDouble() < 0.2f)
-				adjacentState.randomTick(level, pos.relative(direction), randomSource);
-		});
+		FastLeafDecay.decayAdjacentLeaves(level, pos, randomSource);
 	}
 }